+++ t1_initialization/t1_initialization/Timer.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <ctime>
 
 class Timer {
