+++ t1_initialization/t1_initialization/scene.h
@@ -1,0 +1,62 @@
+// CubemMap.h - class for rendering cubemap
+#pragma once
+
+#include <d3dcompiler.h>
+#include <dxgi.h>
+#include <d3d11.h>
+#include <directxmath.h>
+#include <string>
+#include <vector>
+
+#include "timer.h"
+#include "texture.h"
+
+using namespace DirectX;
+
+struct SimpleVertex
+{
+  float x, y, z;      // positional coords
+  float u, v;         // texture coords
+};
+
+struct WorldMatrixBuffer {
+  XMMATRIX worldMatrix;
+};
+
+struct SceneMatrixBuffer {
+  XMMATRIX viewProjectionMatrix;
+};
+
+class Scene {
+public:
+  HRESULT Init(ID3D11Device* device, ID3D11DeviceContext* context, int screenWidth, int screenHeight);
+
+  void Realese();
+
+  void Resize(int screenWidth, int screenHeight) {};
+
+  void Render(ID3D11DeviceContext* context);
+
+  bool Frame(ID3D11DeviceContext* context, XMMATRIX viewMatrix, XMMATRIX projectionMatrix, XMFLOAT3 cameraPos);
+
+private:
+  HRESULT CompileShaderFromFile(const WCHAR* szFileName, LPCSTR szEntryPoint, LPCSTR szShaderModel, ID3DBlob** ppBlobOut);
+
+  // dx11 vars
+  ID3D11VertexShader* g_pVertexShader = nullptr;
+  ID3D11PixelShader* g_pPixelShader = nullptr;
+  ID3D11InputLayout* g_pVertexLayout = nullptr;
+
+  ID3D11Buffer* g_pVertexBuffer = nullptr;
+  ID3D11Buffer* g_pIndexBuffer = nullptr;
+  ID3D11Buffer* g_pWorldMatrixBuffer = nullptr;
+  ID3D11Buffer* g_pSceneMatrixBuffer = nullptr;
+  ID3D11RasterizerState* g_pRasterizerState = nullptr;
+  ID3D11SamplerState* g_pSamplerState = nullptr;
+
+  // Texture with aphex twin
+  Texture txt;
+
+  // Velocity of world matrix rotation
+  float angle_velocity = 3.1415926f;
+};
